non-integer appear times are possible, make tempo macro more intuitve

+++ engine.py
@@ -168,7 +168,7 @@ while not done:
                 for row in csv.reader(level_file):
                     try:
                         if row[0] != 'appear_time' and row[0] != '#' and row[0] != '//':
-                            appear_time = int(row[0])
+                            appear_time = float(row[0])
                             key = int(row[1]) % KEYS
                             instrument = row[2].strip()
                             note = row[3].strip()
